refactor: move compass services into ferriskey-compass
