+++ libs/ferriskey-compass/src/lib.rs
@@ -4,3 +4,4 @@ pub mod recorder;
 pub mod value_objects;
 
 pub mod policies;
+pub mod services;
